move desktop nav array to consts tsx file
